requirements.txt
langchain==0.1.13
openai==1.23.0
requests>=2.25.1
python-dotenv>=1.0.0
faiss-cpu==1.7.4
sentence-transformers==2.5.1
